+++ src/main/java/com/github/hcsp/inheritance/Animal.java
@@ -1,0 +1,13 @@
+package com.github.hcsp.inheritance;
+
+public class Animal {
+    public String name;
+
+    public Animal(String name) {
+        this.name = name;
+    }
+
+    public void sayMyName() {
+        System.out.println("我的名字是" + name);
+    }
+}
