+++ src/main/java/com/github/hcsp/inheritance/Cat.java
@@ -1,14 +1,8 @@
 package com.github.hcsp.inheritance;
 
-public class Cat {
-    private String name;
-
+public class Cat extends Animal{
     public Cat(String name) {
-        this.name = name;
-    }
-
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
+        super(name);
     }
 
     public void meow() {
