+++ src/main/java/com/github/hcsp/inheritance/People.java
@@ -1,0 +1,19 @@
+package com.github.hcsp.inheritance;
+
+public class People {
+    public String name;
+    public int age;
+
+    public People(String name,int age) {
+        this.name = name;
+        this.age = age;
+    }
+
+    public void sayMyName() {
+        System.out.println("我的名字是" + name);
+    }
+
+    public void sayMyAge() {
+        System.out.println("我的年龄是" + age);
+    }
+}
